backend/pyproject.toml
[project]
name = "datanova-agent1"
version = "0.1.0"
description = "Agent IA de veille réglementaire automatisée - Projet PING"
authors = [{name = "PING-DataNova Team"}]
requires-python = ">=3.11"
dependencies = [
    # LangChain & LLM
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    # Web Scraping
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    # PDF Processing
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    # Configuration & Environment
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Logging
    "structlog>=24.4.0",
    # Terminal UI
    "rich>=13.7.0",
    # Email
    "aiosmtplib>=3.0.0",
    # Utilities
    "python-dateutil>=2.9.0",
    "pytz>=2024.1",
    "langgraph>=1.0.5",
    "scrapy>=2.14.1",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "email-validator>=2.3.0",
    "bcrypt>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-timeout>=2.3.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
